+++ src/eyed3/utils/console.py
@@ -133,7 +133,8 @@ class AnsiCodes(object):
             AnsiCodes._USE_ANSI = False
         else:
             AnsiCodes._USE_ANSI = True
-            if "TERM" in os.environ and os.environ["TERM"] == "dumb":
+            if (("TERM" in os.environ and os.environ["TERM"] == "dumb") or
+                ("OS" in os.environ and os.environ["OS"] == "Windows_NT")):
                 AnsiCodes._USE_ANSI = False
 
 
@@ -483,7 +484,11 @@ def _encode(s):
     for python3 (except assert that ``s`` is a unicode type).'''
     if PY2:
         if isinstance(s, unicode):
-            return s.encode(LOCAL_ENCODING)
+            try:
+                return s.encode(LOCAL_ENCODING)
+            except:
+                log.exception()
+                return s.encode(LOCAL_ENCODING, "replace")
         elif isinstance(s, str):
             return s
         else: